pyproject.toml
[tool.poetry]
name = "replica"
version = "0.1.0"
description = ""
authors = ["Gonzalo Ayuso <gonzalo123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
